feat(build): enable -Wunsuffixed-float-constants
Use float suffixes consistently

+++ src/platform-aws.c
@@ -41,7 +41,7 @@ static struct ec2_platform_data platform_data_map[] = {
 		.name = "p4d.24xlarge",
 		.topology = "p4d-24xl-topo.xml",
 		.default_dup_conns = 0,
-		.latency = 75.0,
+		.latency = 75.0f,
 		.gdr_required = true,
 		.net_flush_required = true,
 		.default_protocol = "SENDRECV",
@@ -51,7 +51,7 @@ static struct ec2_platform_data platform_data_map[] = {
 		.name = "p4de.24xlarge",
 		.topology = "p4de-24xl-topo.xml",
 		.default_dup_conns = 0,
-		.latency = 75.0,
+		.latency = 75.0f,
 		.gdr_required = true,
 		.net_flush_required = true,
 		.default_protocol = "SENDRECV",
@@ -61,7 +61,7 @@ static struct ec2_platform_data platform_data_map[] = {
 		.name = "p3dn.24xlarge",
 		.topology = NULL,
 		.default_dup_conns = 4,
-		.latency = 150.0,
+		.latency = 150.0f,
 		.gdr_required = false,
 		.net_flush_required = true,
 		.default_protocol = "SENDRECV",
@@ -71,7 +71,7 @@ static struct ec2_platform_data platform_data_map[] = {
 		.name = "p5.48xlarge",
 		.topology = NULL,
 		.default_dup_conns = 0,
-		.latency = 75.0,
+		.latency = 75.0f,
 		.gdr_required = true,
 		.net_flush_required = false,
 		.default_protocol = "RDMA",
@@ -81,7 +81,7 @@ static struct ec2_platform_data platform_data_map[] = {
 		.name = "p5e.48xlarge",
 		.topology = NULL,
 		.default_dup_conns = 0,
-		.latency = 75.0,
+		.latency = 75.0f,
 		.gdr_required = true,
 		.net_flush_required = false,
 		.default_protocol = "RDMA",
@@ -91,7 +91,7 @@ static struct ec2_platform_data platform_data_map[] = {
 		.name = "g5.48xlarge",
 		.topology = "g5.48xl-topo.xml",
 		.default_dup_conns = 0,
-		.latency = 75.0,
+		.latency = 75.0f,
 		.gdr_required = false,
 		.net_flush_required = true,
 		.default_protocol = "SENDRECV",
@@ -101,7 +101,7 @@ static struct ec2_platform_data platform_data_map[] = {
 		.name = "trn1.32xlarge",
 		.topology = NULL,
 		.default_dup_conns = 0,
-		.latency = 75.0,
+		.latency = 75.0f,
 		.gdr_required = true,
 		.net_flush_required = true,
 		.default_protocol = "SENDRECV",
@@ -111,7 +111,7 @@ static struct ec2_platform_data platform_data_map[] = {
 		.name = "trn1n.32xlarge",
 		.topology = NULL,
 		.default_dup_conns = 0,
-		.latency = 75.0,
+		.latency = 75.0f,
 		.gdr_required = true,
 		.net_flush_required = true,
 		.default_protocol = "SENDRECV",
@@ -121,7 +121,7 @@ static struct ec2_platform_data platform_data_map[] = {
 		.name = "trn2n.48xlarge",
 		.topology = NULL,
 		.default_dup_conns = 0,
-		.latency = 75.0,
+		.latency = 75.0f,
 		.gdr_required = true,
 		.net_flush_required = true,
 		.default_protocol = "RDMA",
@@ -599,14 +599,13 @@ int platform_init(const char **provider_filter)
 		nic_dup_conns = platform_data->default_dup_conns;
 
 	if (ofi_nccl_net_latency() < 0) {
-		if (platform_data && platform_data->latency >= 0.0) {
+		if (platform_data && platform_data->latency >= 0.0f) {
 			net_latency = platform_data->latency;
 		} else {
 			/* For historical reasons, default value for EFA is 150 us */
-			net_latency = 150.0;
+			net_latency = 150.0f;
 		}
-		NCCL_OFI_INFO(NCCL_INIT | NCCL_NET, "Internode latency set at %.1f us",
-				net_latency);
+		NCCL_OFI_INFO(NCCL_INIT | NCCL_NET, "Internode latency set at %.1lf us", (double)net_latency);
 	}
 
 	if (select_efa && ofi_nccl_protocol() == NULL && platform_data) {
